+++ rustls/tests/api.rs
@@ -21,10 +21,6 @@ use rustls::{ServerConfig, ServerConnection};
 use rustls::{Stream, StreamOwned};
 use rustls::{SupportedCipherSuite, ALL_CIPHER_SUITES};
 
-#[cfg(feature = "dangerous_configuration")]
-use rustls::server::ClientCertVerified;
-
-#[allow(dead_code)]
 mod common;
 use crate::common::*;
 
@@ -882,400 +878,6 @@ fn client_auth_works() {
             let (mut client, mut server) =
                 make_pair_for_arc_configs(&Arc::new(client_config), &server_config);
             do_handshake(&mut client, &mut server);
-        }
-    }
-}
-
-#[cfg(feature = "dangerous_configuration")]
-mod test_clientverifier {
-    use super::*;
-    use crate::common::MockClientVerifier;
-    use rustls::internal::msgs::enums::AlertDescription;
-    use rustls::internal::msgs::enums::ContentType;
-
-    // Client is authorized!
-    fn ver_ok() -> Result<ClientCertVerified, Error> {
-        Ok(rustls::server::ClientCertVerified::assertion())
-    }
-
-    // Use when we shouldn't even attempt verification
-    fn ver_unreachable() -> Result<ClientCertVerified, Error> {
-        unreachable!()
-    }
-
-    // Verifier that returns an error that we can expect
-    fn ver_err() -> Result<ClientCertVerified, Error> {
-        Err(Error::General("test err".to_string()))
-    }
-
-    fn server_config_with_verifier(
-        kt: KeyType,
-        client_cert_verifier: MockClientVerifier,
-    ) -> ServerConfig {
-        ServerConfig::builder()
-            .with_safe_defaults()
-            .with_client_cert_verifier(Arc::new(client_cert_verifier))
-            .with_single_cert(kt.get_chain(), kt.get_key())
-            .unwrap()
-    }
-
-    #[test]
-    // Happy path, we resolve to a root, it is verified OK, should be able to connect
-    fn client_verifier_works() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let client_verifier = MockClientVerifier {
-                verified: ver_ok,
-                subjects: Some(get_client_root_store(*kt).subjects()),
-                mandatory: Some(true),
-                offered_schemes: None,
-            };
-
-            let server_config = server_config_with_verifier(*kt, client_verifier);
-            let server_config = Arc::new(server_config);
-
-            for version in rustls::ALL_VERSIONS {
-                let client_config = make_client_config_with_versions_with_auth(*kt, &[version]);
-                let (mut client, mut server) =
-                    make_pair_for_arc_configs(&Arc::new(client_config.clone()), &server_config);
-                let err = do_handshake_until_error(&mut client, &mut server);
-                assert_eq!(err, Ok(()));
-            }
-        }
-    }
-
-    // Server offers no verification schemes
-    #[test]
-    fn client_verifier_no_schemes() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let client_verifier = MockClientVerifier {
-                verified: ver_ok,
-                subjects: Some(get_client_root_store(*kt).subjects()),
-                mandatory: Some(true),
-                offered_schemes: Some(vec![]),
-            };
-
-            let server_config = server_config_with_verifier(*kt, client_verifier);
-            let server_config = Arc::new(server_config);
-
-            for version in rustls::ALL_VERSIONS {
-                let client_config = make_client_config_with_versions_with_auth(*kt, &[version]);
-                let (mut client, mut server) =
-                    make_pair_for_arc_configs(&Arc::new(client_config.clone()), &server_config);
-                let err = do_handshake_until_error(&mut client, &mut server);
-                assert_eq!(
-                    err,
-                    Err(ErrorFromPeer::Client(Error::CorruptMessagePayload(
-                        ContentType::Handshake
-                    )))
-                );
-            }
-        }
-    }
-
-    // Common case, we do not find a root store to resolve to
-    #[test]
-    fn client_verifier_no_root() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let client_verifier = MockClientVerifier {
-                verified: ver_ok,
-                subjects: None,
-                mandatory: Some(true),
-                offered_schemes: None,
-            };
-
-            let server_config = server_config_with_verifier(*kt, client_verifier);
-            let server_config = Arc::new(server_config);
-
-            for version in rustls::ALL_VERSIONS {
-                let client_config = make_client_config_with_versions_with_auth(*kt, &[version]);
-                let mut server = ServerConnection::new(Arc::clone(&server_config)).unwrap();
-                let mut client =
-                    ClientConnection::new(Arc::new(client_config), dns_name("notlocalhost"))
-                        .unwrap();
-                let errs = do_handshake_until_both_error(&mut client, &mut server);
-                assert_eq!(
-                    errs,
-                    Err(vec![
-                        ErrorFromPeer::Server(Error::General(
-                            "client rejected by client_auth_root_subjects".into()
-                        )),
-                        ErrorFromPeer::Client(Error::AlertReceived(AlertDescription::AccessDenied))
-                    ])
-                );
-            }
-        }
-    }
-
-    // If we cannot resolve a root, we cannot decide if auth is mandatory
-    #[test]
-    fn client_verifier_no_auth_no_root() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let client_verifier = MockClientVerifier {
-                verified: ver_unreachable,
-                subjects: None,
-                mandatory: Some(true),
-                offered_schemes: None,
-            };
-
-            let server_config = server_config_with_verifier(*kt, client_verifier);
-            let server_config = Arc::new(server_config);
-
-            for version in rustls::ALL_VERSIONS {
-                let client_config = make_client_config_with_versions(*kt, &[version]);
-                let mut server = ServerConnection::new(Arc::clone(&server_config)).unwrap();
-                let mut client =
-                    ClientConnection::new(Arc::new(client_config), dns_name("notlocalhost"))
-                        .unwrap();
-                let errs = do_handshake_until_both_error(&mut client, &mut server);
-                assert_eq!(
-                    errs,
-                    Err(vec![
-                        ErrorFromPeer::Server(Error::General(
-                            "client rejected by client_auth_root_subjects".into()
-                        )),
-                        ErrorFromPeer::Client(Error::AlertReceived(AlertDescription::AccessDenied))
-                    ])
-                );
-            }
-        }
-    }
-
-    // If we do have a root, we must do auth
-    #[test]
-    fn client_verifier_no_auth_yes_root() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let client_verifier = MockClientVerifier {
-                verified: ver_unreachable,
-                subjects: Some(get_client_root_store(*kt).subjects()),
-                mandatory: Some(true),
-                offered_schemes: None,
-            };
-
-            let server_config = server_config_with_verifier(*kt, client_verifier);
-            let server_config = Arc::new(server_config);
-
-            for version in rustls::ALL_VERSIONS {
-                let client_config = make_client_config_with_versions(*kt, &[version]);
-                let mut server = ServerConnection::new(Arc::clone(&server_config)).unwrap();
-                let mut client =
-                    ClientConnection::new(Arc::new(client_config), dns_name("localhost")).unwrap();
-                let errs = do_handshake_until_both_error(&mut client, &mut server);
-                assert_eq!(
-                    errs,
-                    Err(vec![
-                        ErrorFromPeer::Server(Error::NoCertificatesPresented),
-                        ErrorFromPeer::Client(Error::AlertReceived(
-                            AlertDescription::CertificateRequired
-                        ))
-                    ])
-                );
-            }
-        }
-    }
-
-    #[test]
-    // Triple checks we propagate the rustls::Error through
-    fn client_verifier_fails_properly() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let client_verifier = MockClientVerifier {
-                verified: ver_err,
-                subjects: Some(get_client_root_store(*kt).subjects()),
-                mandatory: Some(true),
-                offered_schemes: None,
-            };
-
-            let server_config = server_config_with_verifier(*kt, client_verifier);
-            let server_config = Arc::new(server_config);
-
-            for version in rustls::ALL_VERSIONS {
-                let client_config = make_client_config_with_versions_with_auth(*kt, &[version]);
-                let mut server = ServerConnection::new(Arc::clone(&server_config)).unwrap();
-                let mut client =
-                    ClientConnection::new(Arc::new(client_config), dns_name("localhost")).unwrap();
-                let err = do_handshake_until_error(&mut client, &mut server);
-                assert_eq!(
-                    err,
-                    Err(ErrorFromPeer::Server(Error::General("test err".into())))
-                );
-            }
-        }
-    }
-
-    #[test]
-    // If a verifier returns a None on Mandatory-ness, then we error out
-    fn client_verifier_must_determine_client_auth_requirement_to_continue() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let client_verifier = MockClientVerifier {
-                verified: ver_ok,
-                subjects: Some(get_client_root_store(*kt).subjects()),
-                mandatory: None,
-                offered_schemes: None,
-            };
-
-            let server_config = server_config_with_verifier(*kt, client_verifier);
-            let server_config = Arc::new(server_config);
-
-            for version in rustls::ALL_VERSIONS {
-                let client_config = make_client_config_with_versions_with_auth(*kt, &[version]);
-                let mut server = ServerConnection::new(Arc::clone(&server_config)).unwrap();
-                let mut client =
-                    ClientConnection::new(Arc::new(client_config), dns_name("localhost")).unwrap();
-                let errs = do_handshake_until_both_error(&mut client, &mut server);
-                assert_eq!(
-                    errs,
-                    Err(vec![
-                        ErrorFromPeer::Server(Error::General(
-                            "client rejected by client_auth_mandatory".into()
-                        )),
-                        ErrorFromPeer::Client(Error::AlertReceived(AlertDescription::AccessDenied))
-                    ])
-                );
-            }
-        }
-    }
-} // mod test_clientverifier
-
-#[cfg(feature = "dangerous_configuration")]
-mod test_serververifier {
-    use super::*;
-    use crate::common::MockServerVerifier;
-    use rustls::internal::msgs::enums::AlertDescription;
-
-    #[test]
-    fn client_can_override_certificate_verification() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let verifier = Arc::new(MockServerVerifier::accepts_anything());
-
-            let server_config = Arc::new(make_server_config(*kt));
-
-            for version in rustls::ALL_VERSIONS {
-                let mut client_config = make_client_config_with_versions(*kt, &[version]);
-                client_config
-                    .dangerous()
-                    .set_certificate_verifier(verifier.clone());
-
-                let (mut client, mut server) =
-                    make_pair_for_arc_configs(&Arc::new(client_config), &server_config);
-                do_handshake(&mut client, &mut server);
-            }
-        }
-    }
-
-    #[test]
-    fn client_can_override_certificate_verification_and_reject_certificate() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let verifier = Arc::new(MockServerVerifier::rejects_certificate(
-                Error::CorruptMessage,
-            ));
-
-            let server_config = Arc::new(make_server_config(*kt));
-
-            for version in rustls::ALL_VERSIONS {
-                let mut client_config = make_client_config_with_versions(*kt, &[version]);
-                client_config
-                    .dangerous()
-                    .set_certificate_verifier(verifier.clone());
-
-                let (mut client, mut server) =
-                    make_pair_for_arc_configs(&Arc::new(client_config), &server_config);
-                let errs = do_handshake_until_both_error(&mut client, &mut server);
-                assert_eq!(
-                    errs,
-                    Err(vec![
-                        ErrorFromPeer::Client(Error::CorruptMessage),
-                        ErrorFromPeer::Server(Error::AlertReceived(
-                            AlertDescription::BadCertificate
-                        ))
-                    ])
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn client_can_override_certificate_verification_and_reject_tls12_signatures() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let mut client_config =
-                make_client_config_with_versions(*kt, &[&rustls::version::TLS12]);
-            let verifier = Arc::new(MockServerVerifier::rejects_tls12_signatures(
-                Error::CorruptMessage,
-            ));
-
-            client_config
-                .dangerous()
-                .set_certificate_verifier(verifier);
-
-            let server_config = Arc::new(make_server_config(*kt));
-
-            let (mut client, mut server) =
-                make_pair_for_arc_configs(&Arc::new(client_config), &server_config);
-            let errs = do_handshake_until_both_error(&mut client, &mut server);
-            assert_eq!(
-                errs,
-                Err(vec![
-                    ErrorFromPeer::Client(Error::CorruptMessage),
-                    ErrorFromPeer::Server(Error::AlertReceived(AlertDescription::BadCertificate))
-                ])
-            );
-        }
-    }
-
-    #[test]
-    fn client_can_override_certificate_verification_and_reject_tls13_signatures() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let mut client_config =
-                make_client_config_with_versions(*kt, &[&rustls::version::TLS13]);
-            let verifier = Arc::new(MockServerVerifier::rejects_tls13_signatures(
-                Error::CorruptMessage,
-            ));
-
-            client_config
-                .dangerous()
-                .set_certificate_verifier(verifier);
-
-            let server_config = Arc::new(make_server_config(*kt));
-
-            let (mut client, mut server) =
-                make_pair_for_arc_configs(&Arc::new(client_config), &server_config);
-            let errs = do_handshake_until_both_error(&mut client, &mut server);
-            assert_eq!(
-                errs,
-                Err(vec![
-                    ErrorFromPeer::Client(Error::CorruptMessage),
-                    ErrorFromPeer::Server(Error::AlertReceived(AlertDescription::BadCertificate))
-                ])
-            );
-        }
-    }
-
-    #[test]
-    fn client_can_override_certificate_verification_and_offer_no_signature_schemes() {
-        for kt in ALL_KEY_TYPES.iter() {
-            let verifier = Arc::new(MockServerVerifier::offers_no_signature_schemes());
-
-            let server_config = Arc::new(make_server_config(*kt));
-
-            for version in rustls::ALL_VERSIONS {
-                let mut client_config = make_client_config_with_versions(*kt, &[version]);
-                client_config
-                    .dangerous()
-                    .set_certificate_verifier(verifier.clone());
-
-                let (mut client, mut server) =
-                    make_pair_for_arc_configs(&Arc::new(client_config), &server_config);
-                let errs = do_handshake_until_both_error(&mut client, &mut server);
-                assert_eq!(
-                    errs,
-                    Err(vec![
-                        ErrorFromPeer::Server(Error::PeerIncompatibleError(
-                            "no overlapping sigschemes".into()
-                        )),
-                        ErrorFromPeer::Client(Error::AlertReceived(
-                            AlertDescription::HandshakeFailure
-                        )),
-                    ])
-                );
-            }
         }
     }
 }
