+++ rustls/tests/common/mod.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use std::convert::{TryFrom, TryInto};
 use std::io;
 use std::ops::{Deref, DerefMut};
@@ -12,17 +14,6 @@ use rustls::RootCertStore;
 use rustls::{Certificate, PrivateKey};
 use rustls::{ClientConfig, ClientConnection};
 use rustls::{ConnectionCommon, ServerConfig, ServerConnection, SideData};
-
-#[cfg(feature = "dangerous_configuration")]
-use rustls::client::{
-    HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier, WebPkiVerifier,
-};
-#[cfg(feature = "dangerous_configuration")]
-use rustls::server::{ClientCertVerified, ClientCertVerifier};
-#[cfg(feature = "dangerous_configuration")]
-use rustls::{
-    internal::msgs::handshake::DigitallySignedStruct, DistinguishedNames, SignatureScheme,
-};
 
 macro_rules! embed_files {
     (
@@ -395,176 +386,6 @@ pub fn do_handshake(
         client.process_new_packets().unwrap();
     }
     (to_server, to_client)
-}
-
-#[cfg(feature = "dangerous_configuration")]
-pub struct MockServerVerifier {
-    cert_rejection_error: Option<Error>,
-    tls12_signature_error: Option<Error>,
-    tls13_signature_error: Option<Error>,
-    wants_scts: bool,
-    signature_schemes: Vec<SignatureScheme>,
-}
-
-#[cfg(feature = "dangerous_configuration")]
-impl ServerCertVerifier for MockServerVerifier {
-    fn verify_server_cert(
-        &self,
-        end_entity: &rustls::Certificate,
-        intermediates: &[rustls::Certificate],
-        server_name: &rustls::ServerName,
-        scts: &mut dyn Iterator<Item = &[u8]>,
-        oscp_response: &[u8],
-        now: std::time::SystemTime,
-    ) -> Result<ServerCertVerified, Error> {
-        let scts: Vec<Vec<u8>> = scts.map(|x| x.to_owned()).collect();
-        println!(
-            "verify_server_cert({:?}, {:?}, {:?}, {:?}, {:?}, {:?})",
-            end_entity, intermediates, server_name, scts, oscp_response, now
-        );
-        if let Some(error) = &self.cert_rejection_error {
-            Err(error.clone())
-        } else {
-            Ok(ServerCertVerified::assertion())
-        }
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        message: &[u8],
-        cert: &Certificate,
-        dss: &DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, Error> {
-        println!(
-            "verify_tls12_signature({:?}, {:?}, {:?})",
-            message, cert, dss
-        );
-        if let Some(error) = &self.tls12_signature_error {
-            Err(error.clone())
-        } else {
-            Ok(HandshakeSignatureValid::assertion())
-        }
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        message: &[u8],
-        cert: &Certificate,
-        dss: &DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, Error> {
-        println!(
-            "verify_tls13_signature({:?}, {:?}, {:?})",
-            message, cert, dss
-        );
-        if let Some(error) = &self.tls13_signature_error {
-            Err(error.clone())
-        } else {
-            Ok(HandshakeSignatureValid::assertion())
-        }
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<SignatureScheme> {
-        self.signature_schemes.clone()
-    }
-
-    fn request_scts(&self) -> bool {
-        println!("request_scts? {:?}", self.wants_scts);
-        self.wants_scts
-    }
-}
-
-#[cfg(feature = "dangerous_configuration")]
-impl MockServerVerifier {
-    pub fn accepts_anything() -> Self {
-        MockServerVerifier {
-            cert_rejection_error: None,
-            ..Default::default()
-        }
-    }
-
-    pub fn requests_scts() -> Self {
-        MockServerVerifier {
-            wants_scts: true,
-            ..Default::default()
-        }
-    }
-
-    pub fn rejects_certificate(err: Error) -> Self {
-        MockServerVerifier {
-            cert_rejection_error: Some(err),
-            ..Default::default()
-        }
-    }
-
-    pub fn rejects_tls12_signatures(err: Error) -> Self {
-        MockServerVerifier {
-            tls12_signature_error: Some(err),
-            ..Default::default()
-        }
-    }
-
-    pub fn rejects_tls13_signatures(err: Error) -> Self {
-        MockServerVerifier {
-            tls13_signature_error: Some(err),
-            ..Default::default()
-        }
-    }
-
-    pub fn offers_no_signature_schemes() -> Self {
-        MockServerVerifier {
-            signature_schemes: vec![],
-            ..Default::default()
-        }
-    }
-}
-
-#[cfg(feature = "dangerous_configuration")]
-impl Default for MockServerVerifier {
-    fn default() -> Self {
-        MockServerVerifier {
-            cert_rejection_error: None,
-            tls12_signature_error: None,
-            tls13_signature_error: None,
-            wants_scts: false,
-            signature_schemes: WebPkiVerifier::verification_schemes(),
-        }
-    }
-}
-
-#[cfg(feature = "dangerous_configuration")]
-pub struct MockClientVerifier {
-    pub verified: fn() -> Result<ClientCertVerified, Error>,
-    pub subjects: Option<DistinguishedNames>,
-    pub mandatory: Option<bool>,
-    pub offered_schemes: Option<Vec<SignatureScheme>>,
-}
-
-#[cfg(feature = "dangerous_configuration")]
-impl ClientCertVerifier for MockClientVerifier {
-    fn client_auth_mandatory(&self) -> Option<bool> {
-        self.mandatory
-    }
-
-    fn client_auth_root_subjects(&self) -> Option<DistinguishedNames> {
-        self.subjects.as_ref().cloned()
-    }
-
-    fn verify_client_cert(
-        &self,
-        _end_entity: &Certificate,
-        _intermediates: &[Certificate],
-        _now: std::time::SystemTime,
-    ) -> Result<ClientCertVerified, Error> {
-        (self.verified)()
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<SignatureScheme> {
-        if let Some(schemes) = &self.offered_schemes {
-            schemes.clone()
-        } else {
-            WebPkiVerifier::verification_schemes()
-        }
-    }
 }
 
 #[derive(PartialEq, Debug)]
